requirements.txt
faiss_cpu==1.9.0.post1
langchain_google_genai==2.0.9
numpy==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.41.1
